btc_feed_aggregator/requirements.txt
flask==2.1.3
Werkzeug==2.1.2
flask_restx==0.5.1
grpcio==1.49.0
grpcio-tools==1.48.1
